commit #3
in the nv, on line 11, A static list (MyItems) to store all UserItems objects.
static: Shared across all instances of the class.
private: Accessible only within the UserItems class.
Initialized as an empty ArrayList, on line 27-28, A static method to add a UserItems object to the static list MyItems. on line 31-39, we were testing our items if available or no